refactor: insert, return money integration & add return money logci
